% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_icon_choices.R
\name{get_icon_choice}
\alias{get_icon_choice}
\title{Get Initial and Final Rates of Icon Choices}
\usage{
get_icon_choice(choiceData, choice_time)
}
\arguments{
\item{choiceData}{The output of \code{\link{readChoices}} on one or more
subjects.}

\item{choice_time}{Either "initial" or "final", representing which segment of
the IAPS choice data we should access.}
}
\value{
A data frame consisting of 36 rows, and six columns corresponding to
  the round number, the presented icons, the numeric choice (option), the
  name of the chosen icon (choice), and the valence group of the chosen icon.
}
\description{
Get Initial and Final Rates of Icon Choices
}
\examples{

get_icon_choice(sampleChoiceData, "initial")

get_icon_choice(sampleChoiceData, "final")

}
