% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getGroupInfo}
\alias{getGroupInfo}
\title{Get Group Reward and Probability Info for All Phases}
\usage{
getGroupInfo(data)
}
\arguments{
\item{data}{The data output from \code{\link{readChoices}}}
}
\value{
A dataframe with 15 rows and four columns. It has the reward and
  probability of receiving it as a function of each image's group per phase.
  This can be compared to the output of \code{\link{getChoices}} or
  \code{\link{processChoiceData}}.
}
\description{
Get Group Reward and Probability Info for All Phases
}
