% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getGroupRewardProbs}
\alias{getGroupRewardProbs}
\title{Get One Phase's Group Reward and Probability Info}
\usage{
getGroupRewardProbs(phaseData)
}
\arguments{
\item{phaseData}{One of the three phase objects defined by calling
\code{\link{getPhases}}.}
}
\value{
A dataframe with four columns. It has the reward and probability of
  receiving it as a function of each image's group per phase. This can be
  compared to the output of \code{\link{getChoices}} or
  \code{\link{processChoiceData}}. This function is called by
  \code{\link{getGroupInfo}} which combines the information for all three of
  the phases.
}
\description{
Get One Phase's Group Reward and Probability Info
}
