% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconPrefs.R
\name{getIconDiffs}
\alias{getIconDiffs}
\title{Get the difference in icon selection}
\usage{
getIconDiffs(choiceData, side = FALSE)
}
\arguments{
\item{choiceData}{The output of \code{\link{readChoices}}.}

\item{side}{Logical: TRUE and the difference will be categorized by the side
the icon appeared on the screen. FALSE and it won't be. Default is FALSE.}
}
\value{
A data frame with three columns containing subject, the icons they
  saw, and the difference between their initial preference for a given icon
  (as a proportion of the times they chose it when available) and their final
  icon preference.
}
\description{
This function gets the difference in the proportions of icon selection during
the initial and final rounds for each subject with a unique combination of
icons.
}
