% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconPrefs.R
\name{getIconOptions}
\alias{getIconOptions}
\title{Which icons were presented?}
\usage{
getIconOptions(df)
}
\arguments{
\item{df}{The iconChoices data frame as output by the \code{\link{getIconPrefs}} function.}
}
\value{
A list with the different icons an individual had to choose from.
}
\description{
This function returns the unique values of an icon for a given subject.
}
