% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconPrefs.R
\name{getIconPrefs}
\alias{getIconPrefs}
\title{Get Icon Preferences Before and After Choice Task}
\usage{
getIconPrefs(choiceData)
}
\arguments{
\item{choiceData}{The output of \code{\link{readChoices}} on one or
more subjects.}
}
\value{
A data frame consisting of 72 rows and 10 columns showing the round
  number, icon options and choice. Additionally, there is a survey column
  which specifies whether or not the icon preference was before or after the
  choice task.
}
\description{
This function calls \code{\link{getPhases}} and uses the objects
phase$initIcon and phase$finalIcon to determine proportion of icons an
individual chose.
}
