% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getIcons}
\alias{getIcons}
\title{Get Icons for Each Round}
\usage{
getIcons(phaseData, pivot = FALSE)
}
\arguments{
\item{phaseData}{One of the three phase objects defined by calling \code{\link{getPhases}}.}

\item{pivot}{Logical: pivoting the data makes it tidier, but makes it harder
to join with the other information. For this reason, it is being set to
FALSE by default.}
}
\value{
Returns a dataframe with four columns:
\itemize{
\item \emph{icon1:} The icon presented in the first order position.
\item \emph{icon2:} The icon presented in the second order position.
\item \emph{phase:} The phase these icons were displayed in (constant for each \code{phase} object).
\item \emph{round:} The round number these icons were displayed in, specific to the phase.
}
}
\description{
This gets the icons for each round in each phase.
}
\examples{
\dontrun{getIcons(phase$one)}
}
