% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getImageShown}
\alias{getImageShown}
\title{Get Images Shown for Each Round}
\usage{
getImageShown(phaseData)
}
\arguments{
\item{phaseData}{One of the three phase objects defined by calling \code{\link{getPhases}}.

#' @return Returns a dataframe with three columns:
\itemize{
\item \emph{image:} The image the subject was shown after making a choice.
\item \emph{phase:} The phase the choices were made in (constant for each \code{phase} object).
\item \emph{round:} The round number these choices were made in, specific to the phase.
}}
}
\description{
This function gets the images shown for each round in each phase.
}
\examples{

\dontrun{getImageShown(phase$three)}
}
