% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getPhaseIcons}
\alias{getPhaseIcons}
\title{Get Icons for Each Round per Phase}
\usage{
getPhaseIcons(phaseData)
}
\arguments{
\item{phaseData}{One of the three phase objects defined by calling
\code{\link{getPhases}}.}
}
\value{
This returns a dataframe with four columns: \itemize{\item
  \emph{phase:} The phase these icons were displayed in (constant for each
  \code{phase} object). \item \emph{round:} The round number these icons were
  displayed in, specific to the phase. \item \emph{icon1:} The icon presented
  in the first order position. \item \emph{icon2:} The icon presented in the
  second order position. }
}
\description{
This function gets the icons for each round in each phase. It \strong{was}
the old \code{\link{getIcons}} function, but is now used in that.
}
