% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{getPhases}
\alias{getPhases}
\title{Get Task Phases}
\usage{
getPhases(data)
}
\arguments{
\item{data}{The data output from \code{\link{readChoices}}}
}
\description{
This function takes the data column from the output of
\code{\link{readChoices}} and separates the three phases of the task, saving
them to the environment \code{phrase}, which is initialized behind the
scenes. This function \strong{must} be run in order to use the other
functions that parse the choice data. If it is not, the phases will need to
be isolated by hand and assigned to the phase environment (or local/global
variables). See the \emph{Details} section for more how to access the three
phases.
}
\details{
The three phases of the task may be accessed by calling the variables:
\itemize{ \item \code{phase$one} \item \code{phase$two} \item
\code{phase$three} }
}
