% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{getPicID}
\alias{getPicID}
\title{Get Picture IDs}
\usage{
getPicID(data)
}
\arguments{
\item{data}{The data output from \code{\link{readRatings}}.}
}
\value{
A dataframe/tibble column titled "picture" with the different images
  displayed on each round.
}
\description{
This function takes the data column from the output of
\code{\link{readRatings}} and filters it for rows that contain the phrase
"Image", and removes the "ROUND 0" trial. It then parses the string into the
different components, isolating the picture ID variable.
}
