% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositive.R
\name{getPositive}
\alias{getPositive}
\title{Filter for positive pictures}
\usage{
getPositive(data)
}
\arguments{
\item{data}{The \code{IAPS_id} file.}
}
\value{
filtered data set
}
\description{
Filter the IAPS_id dataset for questions where individuals were asked "How positive does this make you feel?"
}
