% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{getQuestion}
\alias{getQuestion}
\title{Get Question Type Asked}
\usage{
getQuestion(data)
}
\arguments{
\item{data}{The data output from \code{\link{readRatings}}.}
}
\value{
A dataframe/tibble column titled "question" with variables positive
  and negative signifying the question asked on each round.
}
\description{
This function takes the data column from the output of
\code{\link{readRatings}} and filters it for rows that contain the phrase
"Question". It then replaces any rows that contain "NEGATIVE," meaning the
subject was asked "How negative does this image make you feel?" and replaces
it with the string "negative." It replaces rows that contain the string
"POSITIVE" with the string "positive."
}
