% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{getRating}
\alias{getRating}
\title{Get Image Ratings}
\usage{
getRating(data)
}
\arguments{
\item{data}{The data output from \code{\link{readRatings}}.}
}
\value{
A dataframe/tibble column titled "rating" with the different images
  displayed on each round.
}
\description{
This function takes the data column from the output of
\code{\link{readRatings}} and filters it for rows that contain the phrase
"SUBMIT", and removes the "ROUND 0" trial. It then parses the strings in each row into the
different components, isolating the submitted rating value variable.
}
