% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IAPS_id}
\alias{IAPS_id}
\title{IAPS IDs - Sample Data}
\format{
A data frame with 70 rows and 3 variables:
\describe{
  \item{picture}{image description with ".P" or ".N" signifying the accompanying question.}
  \item{IAPS_ID}{image ID in IAPS database}
  \item{googleID}{image ID in Google Survey}
  ...
}
}
\source{
IAPS dataset
}
\usage{
IAPS_id
}
\description{
A dataset containing the descriptions and IAPS/Google ID numbers for a subset of the IAPS dataset.
}
\keyword{datasets}
