% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iapsr-package.R
\docType{package}
\name{iapsr-package}
\alias{iapsr}
\alias{iapsr-package}
\title{iapsr: IAPS Analysis Functions}
\description{
Package useful for analyzing images as part of the International Affective Picture System.
}
\keyword{internal}
