% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imageGroupings}
\alias{imageGroupings}
\title{Images in each group -- PRPT Scale}
\format{
A data frame with 190 rows and 2 columns
\describe{
  \item{IAPS}{The IAPS ID.}
  \item{group}{The PRPT group to which the image was mapped.}
}
}
\source{
Kishida Lab
}
\usage{
imageGroupings
}
\description{
A dataset containing the images used in each of the PRPT groups that the IAPS Choice task is modeled upon.
}
\keyword{datasets}
