% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_rows.R
\name{mutate_rows}
\alias{mutate_rows}
\title{Mutate at specific rows based on a specific condition.}
\usage{
mutate_rows(.data, .predicate, ...)
}
\arguments{
\item{.data}{The data being supplioed.}

\item{.predicate}{The condition to be met}

\item{...}{Any number of columns/values to impute.}
}
\value{
The mutated data frame.
}
\description{
This function was developed by Davis Vaughan and can be found at his GitHub
gist. https://gist.github.com/DavisVaughan/24cbc404c09e75d3bf23467d15a7d42d.
}
\details{
I am using it in this package as it seems more logical to me than to mutate by column and
case_when
}
