% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconPrefs.R
\name{organizeIconGroupings}
\alias{organizeIconGroupings}
\title{Organize Icon Group Mappings for Every Subject}
\usage{
organizeIconGroupings(rawChoiceDF, processedRatingDF)
}
\arguments{
\item{rawChoiceDF}{The output of \code{\link{readChoices}}.}

\item{processedRatingDF}{The output of \code{\link{processRatingsData}}.}
}
\value{
A dataframe with the subject-specific mean rating for each icon class
  (corresponding to the PRPT task) as well as the EU and difference in
  preference over time.
}
\description{
This function maps the group-icon information for each subject with the
difference in their icon preference from before the choice task and after. It
calls the functions \code{\link{getGroupInfo}} and
\code{\link{getIconDiffs}}.
}
\examples{

#RJTRJTRJT is the subject name for this example. It is set up this in the example
#since the function is built for bactch processing and parses the subject ID out
#of the file path which is nine characters.
choice <- base::list("Path File/choice/RJTRJTRJTchoice.txt.txt" = sampleChoiceData)
rate <- sampleRatingsData \%>\%
 processRatingsData() \%>\%
 #add the picture 7006 to prevent error since new cohort didn't rate it.
 dplyr::add_row(round = 00, picture = "7006",
                question = "negative", rating = 1.73913) \%>\%
 dplyr::add_row(round = 01, picture = "7006",
                question = "positive", rating = 2.630435) \%>\%
 regSetup() \%>\%
 dplyr::mutate(subject = "RJTRJTRJT", .before = IAPS)
organizeIconGroupings(choice, rate)

}
