% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examineChoices.R
\name{plotExamined}
\alias{plotExamined}
\title{Plot the Percent of Optimal Choices for Each Subject}
\usage{
plotExamined(processedData, facet = FALSE, subjectName = NULL)
}
\arguments{
\item{processedData}{The output of \code{\link{processChoiceData}}.}

\item{facet}{Logical: TRUE and plots will be faceted by subject. FALSE and it
won't. Default is FALSE.}

\item{subjectName}{Optional argument to include a subject name if plotting
for only one subject. Name must be in quotations.}
}
\value{
A plot showing the percent of optimal choices over time (round
  number), with each phase specified.
}
\description{
This function is an adapted form of \code{\link{plotPercentOptimal}}. It
takes the output of \code{\link{examineChoices}}. It checks to see if the
data has been modified to include runningRound and percentOptimal and, if
not, does so. See examples for more detail.
}
\examples{

choice <- processChoiceData(sampleChoiceData) \%>\%
dplyr::mutate(subject = "RJT", .before = phase)

rate <- processRatingsData(sampleRatingsData) \%>\%
regSetup() \%>\%
dplyr::mutate(subject = "RJT", .before = IAPS)

# The plotExamined can be used with the output of examineChoices that is
# modified:
examineChoices(choice, rate) \%>\%
 dplyr::group_by(subject, phase) \%>\%
 dplyr::mutate(percentOptimal = base::cumsum(optimal) / dplyr::row_number()) \%>\%
 dplyr::ungroup() \%>\%
 dplyr::group_by(subject) \%>\%
 dplyr::mutate(runningRound = dplyr::row_number()) \%>\%
 dplyr::ungroup() \%>\% plotExamined()

# This function can be used with the output of examineChoices that is not
# modified and will modify it internally:

examineChoices(choice, rate) \%>\% plotExamined()


}
