% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{plotPercentOptimal}
\alias{plotPercentOptimal}
\title{Plot the Percent of Optimal Choices for Each Subject}
\usage{
plotPercentOptimal(processedData, facet = FALSE)
}
\arguments{
\item{processedData}{The output of \code{\link{processChoiceData}}.}

\item{facet}{Logical: TRUE and plots will be faceted by subject. FALSE and it won't. Default is FALSE.}
}
\value{
A plot showing the percent of optimal choices over time (round number), with each phase specified.
}
\description{
Plot the Percent of Optimal Choices for Each Subject
}
