% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{processChoiceData}
\alias{processChoiceData}
\title{Process Choice Data}
\usage{
processChoiceData(data)
}
\arguments{
\item{data}{The data output from \code{\link{readChoices}}}
}
\value{
A dataframe 150 rows and 8 columns: \itemize{ \item \emph{phase:} The
  phase the choices were made in. \item \emph{round:} The round number these
  choices were made in, specific to the phase. \item \emph{icon1:} The icon
  presented in the first order position. \item \emph{icon2:} The icon
  presented in the second order position. \item \emph{option:} The option the
  subject chose. This should correspond to the icon order as presented in the
  task. \item \emph{choice:} The icon the subject chose. \item \emph{group:}
  The (payment weighting) group the icon corresponds to. \item \emph{image:}
  The image the subject was shown after making a choice. }
}
\description{
This function takes the data column from the output of
\code{\link{readChoices}} and runs four functions, binding the results in a
dataframe for easy manipulation: \itemize{ \item \code{\link{getPhases}}
which separates the three phases of the task, saving them to the environment
\code{phrase}, which is initialized behind the scenes. \item
\code{\link{getIcons}} which gets the icon options shown to the subject.
\item \code{\link{getChoices}} which gets the choice the subject makes. \item
\code{\link{getImageShown}} which gets the images the subject is shown after
choosing an icon.}
}
