% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{processRatingsData}
\alias{processRatingsData}
\title{Process Ratings Data}
\usage{
processRatingsData(data)
}
\arguments{
\item{data}{The data output from \code{\link{readRatings}}.}
}
\value{
Dataframe/tibble with 242 rows and four columns containing information
 about the round, picture ID, question asked, and rating value. See description
 for more information.
}
\description{
This function takes the data column from the output of
\code{\link{readRatings}} and calls the \code{\link{getPicID}},
\code{\link{getQuestion}}, and \code{\link{getRating}} functions. It combines
the outputs of them, and returns a dataframe/tibble with 242 rows and four
columns: \itemize{ \item \emph{round:} The round number (integer) \item \emph{picture:}
The picture ID from IAPS dataset (factor) \item \emph{question:} "positive" or
"negative" values based on whether the question was "How positive does this
make you feel?" or "How negative does this make you feel?" (factor) \item
\emph{rating:} The image rating for that round/question combination (double)}
}
