% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{readChoices}
\alias{readChoices}
\title{Read Choice Data}
\usage{
readChoices(filePath)
}
\arguments{
\item{filePath}{The path to the ratings data file. Should be in single quotes.}
}
\value{
The raw data downloaded from the task. It is a data frame with two columns, time and data, and 1875 observations.
}
\description{
Read Choice Data
}
