% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{readRatings}
\alias{readRatings}
\title{Read Ratings Data}
\usage{
readRatings(filePath)
}
\arguments{
\item{filePath}{The path to the ratings data file. Should be in single quotes.}
}
\value{
The raw data downloaded from the task. It is a data frame with two columns, time and data, and 3631 observations.
}
\description{
Read Ratings Data
}
