% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRatings.R
\name{regSetup}
\alias{regSetup}
\title{Setup Data for Testing Regression Models}
\usage{
regSetup(processedData)
}
\arguments{
\item{processedData}{The data output from \code{\link{processRatingsData}}.}
}
\value{
A tibble that's 120 rows by 3 columns. It has the IAPS column, which
  is a factor that contains the IDs of each image in the IAPS data set, as
  well as a negative and positive column that correspond to the ratings for
  each image when the subjects were asked "How positive does this image make
  you feel?" and "How negative does this image make you feel?"
}
\description{
This function is to be used in order to set up the new data collected from
the IAPS task to test the regression models generated from previous data. It
can be used in batch formulation, joining by subject in addition to picture
rating.
}
