% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleChoiceData}
\alias{sampleChoiceData}
\title{Sample Data from the IAPS Choice Task}
\format{
A data frame with 1863 rows and 2 columns
\describe{
  \item{time}{Time stamps for choice task.}
  \item{data}{data from the choice task}
}
}
\source{
Kishida Lab (researcher)
}
\usage{
sampleChoiceData
}
\description{
A dataset containing output from the IAPS choice task. This was performed by
a study team member. This object has been read in using the
\code{\link{readChoices}} function.
}
\keyword{datasets}
