% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleRatingsData}
\alias{sampleRatingsData}
\title{Sample Data from the IAPS Ratings Task}
\format{
A data frame with 3640 rows and 2 columns
\describe{
  \item{time}{Time stamps for ratings task.}
  \item{data}{data from the ratings task}
}
}
\source{
Kishida Lab (researcher)
}
\usage{
sampleRatingsData
}
\description{
A dataset containing output from the IAPS ratings task. This was performed by
a study team member. This object has been read in using the
\code{\link{readRatings}} function.
}
\keyword{datasets}
