% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconPrefs.R
\name{summarizeIconChoices}
\alias{summarizeIconChoices}
\title{Summarize Icon Choices}
\usage{
summarizeIconChoices(df, icon, type)
}
\arguments{
\item{df}{The iconChoices data frame as output from
\code{\link{getIconPrefs}}.}

\item{icon}{One of the icons in the list as output by the function
\code{\link{getIconOptions}}.}

\item{type}{A character string "initial" or "final" signifying which icon
choice summary to get.}
}
\value{
A data frame with two rows and three columns showing the subject, the
  side the icon was displayed on, and the proportion the image was chosen
  during either the initial survey or final survey (depending on the type).
}
\description{
This function gets the initial or final proportion of times an icon is chosen
by an individual.
}
