% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{Sample Data from the IAPS Choice Task}
\format{
A data frame with 1863 rows and 2 columns
\describe{
  \item{time}{Time stamps for choice task.}
  \item{data}{data from the choice task}
  ...
}
}
\source{
IAPS dataset
}
\usage{
test
}
\description{
A dataset containing the a sample output from the IAPS choice task.
}
\keyword{datasets}
