% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChoices.R
\name{processChoiceData}
\alias{processChoiceData}
\title{Process Choice Data}
\usage{
processChoiceData(data)
}
\arguments{
\item{data}{The data output from \code{\link{readChoices}}}
}
\value{
A dataframe 150 rows and 13 columns: \itemize{ \item
  \emph{phase:} The phase these icons were displayed in (constant for each
  \code{phase} object). \item \emph{round:} The round number these icons were
  displayed in, specific to the phase. \item \emph{icon1:} The icon presented
  in the first order position. \item \emph{icon1Group:} The group mapping for
  the icon presented in the first order position. \item \emph{icon1Prob:} The
  probability of getting a reinforcing image by choosing icon 1. \item
  \emph{icon1Sign:} The relative sign of the reinforcing image (positive or
  negative) by choosing icon 1. \item \emph{icon2:} The icon presented in the
  second order position. \item \emph{icon2Group:} The group mapping for the
  icon presented in the second order position. \item \emph{icon2Prob:} The
  probability of getting a reinforcing image by choosing icon 2. \item
  \emph{icon2Sign:} The relative sign of the reinforcing image (positive or
  negative) by choosing icon 2. \item \emph{chosenIcon:} The icon the subject
  chose. \item \emph{chosenIconGroup:} The (payment weighting) group the
  chosen icon corresponds to. \item \emph{reinforcer:} The image the subject
  was shown (reinforced with) after making a choice.}
}
\description{
This function takes the data column from the output of
\code{\link{readChoices}} and runs four functions, binding the results in a
dataframe for easy manipulation: \itemize{ \item \code{\link{getPhases}}
which separates the three phases of the task, saving them to the environment
\code{phrase}, which is initialized behind the scenes. \item
\code{\link{getIcons}} which gets the icon information shown to the subject
per round. \item \code{\link{getChoices}} which gets the choice the subject
makes. \item \code{\link{getImageShown}} which gets the images the subject is
shown after choosing an icon.}
}
